requirements.txt
customtkinter==5.2.2
matplotlib==3.10.3
nltk==3.9.1
numpy==1.23.5
pandas==2.2.3
Pillow==11.2.1
pycountry==24.6.1
pyterrier==0.1.5
python_terrier==0.13.0
Requests==2.32.3
seaborn==0.13.2
sentence_transformers==4.1.0
python==3.11